Add variable to hold eventual returned data
This variable will be given data when the Search button is clicked and results in a resolved promise object.
The buildToday function will take in that data and pluck from it the current days weather data to then fill the #current-weather section elements children elements with data.
Converted this file from just holding variables and a function to holding a module pattern, which I don't think I've correctly exported yet.