Move variables into event listener where needed
Beginners error I had a variable pull the input fields entered value at load and not when the button is clicked, which resulted in moving other variables into that event listener function as well.
I also changed that event listener function to be async so we could await the fetch response.