Add code for getWeather, declare new function
In addition to adding variables that target specific pre made elements, I've declared a function that will pick from a set of imported images to be the background image for a days forecast, be it current day or future days.
I have yet to import nor even find such images but wanted to get the pseudocode written for later.
I've also added some pseudocode to describe some data that dynamic elements will need to be made for, for displaying data.