Change images to jpg, add pickTextColor
The gifs required more bandwidth and styling for the text to be readable so replaced them with jpg static images.
The newly added pickTextColor is a function that changes the color of the font inside the current weather section depending on the weather (what image is used), the styling isn't great and I'll come back to it later when my mvp is done... I still need to work on the next 7 days of weather.
I also changed around some terms and created a variable that holds the backgroundImage as we needed to share that value with pickTextColor.