Add pseudo code comment for conversion functions
I wanted to note an idea I had for later about the functions that convert data, maybe I can call these functions from the start to default add MPH or F to the end, and then when given an argument with a F or trailing C convert and return to the opposite, similar operations to the other function.
I'm still contemplating how to convert inline dynamically created content, I could add an ID value to these convert able fields or I could make another entire request which I'd rather not do... So I'll aim for adding id values and tying them to the iterator which should related them to the trailing id value integer of their section and then later we'll use the id values to convert the value and replace the text content somehow.