Add conversion functions, manipulate response
I added some more variables for targeting pre made areas and to store data, the futureWeatherHeading was needed because I had to hide the elements content on page load as saying "Weather over the next week" without any data being shown is annoying to look at, and returnedData will hold the converted response data so that we can manipulate it to have Imperial data lengths added which was needed so later we can convert to metric units and back and forth.
Added code in buildToday to delete any prior dynamically produced code so if a user wants to look up another locations weather without refreshing their page now they are able to do so.
I have removed the preexisting data units from the dynamically created content in the buildToday, as we take care of that elsewhere, more efficiently.
Added code to fillNextSeven that checks whether there is existing content in the future day sections and if so removes it.
Added logic for convertButton listener, which allows us to swap data units and convert the data to the appropriate value, this function also calls buildToday and fillNextSeven resulting in a repaint, and I'm sure theres a way to just change the values in the preexisting elements yet I picked this instead because I knew how to do it and would like to mover forward.
Added conversion functions that take in specific values and convert to the opposite (imperial to metric and vice versa), these are used inside the listener on the convert button.