Add js validation for input field, loadingModal
The input field has a minlength of 3 validation attached to it but it wasn't being enforced by either the html nor my js code, by adding an onSubmit of return false and removing the e.preventDefault inside the getWeatherButton listener code we now allow html validation to work and tell the user that their input is not long enough when it's less than 3 and inside our js we also don't make a request until the input length is 3 or more characters.
There is still problems with requesting non cities and it returning an error so I need to address that.
I also added two functions, one for creating a loading message to users while their request is being processed and another for hiding the message when their request's response is received and processed.
These are not working correctly on new input entry, nor have I used them during conversion, however conversion is not an async operation so I don't think I will use it there.