Add required to input, onSubmit to form
The onSubmit replaces are preventDefault of the buttons submit action on the form data as we don't want to refresh the page, however this way allows us to show users the constraint validation api error messages in real time versus adding our own.
I had forgotten to previously use the required attribute on the input element which is needed for the validation to happen.