Add requirements for conversion, error mitigation
If the convert button is clicked before any query has been made and response data sent back then this button would have caused problems, however now clicking this button without first querying for information from the api results in zero actions.
I also added a conditional that if the response object has an error property than we will alert the user, for now it alerts them that their input was bad whenever there is an error returned, which we used the catch method to send back an object with a property of error linking to the error the promise rejected with in the weatherOperations file.
In addition I added an error message to be displayed under the input field so the users are alerted to change their input values and try again, and to do this I needed to create another variable that targeted the form element so we could temporarily add our span element error message after the form element, which is displayed for 3 seconds and then removed.
Added some pseudo code to remind me that the error message is not specific to the users fault at this time.