Change placeholder on input, add id to form
I've learned this api is really good and can destruct a length string to look for specific locations not just cities so have changed the placeholder value to just tell the user 3 characters are needed at minimum.
I added the id value for the form so I could target the form in my js when I need to display an error message regarding input to the user.