Style loading modal and error message
The loading modal was not positioned in a easy to find position when a pre existing locations weather data was displayed, now it is positioned absolutely on top of previous data so users can see that the application is working behind the scenes.
I had to style the error message so that It was not only underneath the input field but sticks out, thus using a red font color and a bigger size.