Change import of images into lazy loaded dynamic
The project guidelines suggested to use the dynamic import() method instead of just importing the images atop our file, this is so that we can lazy load them and reduce the bandwidth needed on loading the app, and these images will be downloaded when they are used.
There are 2 problems here: I think I have written this correctly so that it pauses rendering until the images are loaded in, and what I have done is not specifically what the project guidelines asked for, I instead think they wanted me to use the dynamic import to bundle all images into one import, so I'll look into that in the future, although the more I think about it the more it doesn't make sense. Why would I import all images for the project when I need only one right now?