Revert image importing to esm version
After many tries with the dynamic import method I've decided to revert to using the esm import practice, which is not lazy loading but the images are so small at this point in size that it shouldn't matter much.
Added a comment to describe what one line of code was doing, and removed comments from pending fixes that I have deemed complete.