Change images for better images
The images I had for a rainy day and a cloudy day were too similar to each other so decided to change the cloudy image, and the sunny day when used as a background for the next 6 days wasn't quite fitting with white color text so instead of having the text change to black font I changed the image. I also thought that having two different font colors could be an annoyance to users especially if each day was inverse of the past font color.